Holy smokes! Lots of stuff!
Lots of stuff here! But bear with me:

Tiles can now be dragged onto the grids which accept their value (an
attribute called action)  and store them in a 2d array

*Any tile dropped outside of the grids is deleted from the total list of
tiles
*Tiles dropped on the grid snap into place
*Tiles make note of where they're dropped and on which graph
*Any tile dragged onto another tile already on a grid overwrites it and
deletes the old tile from the list of total tiles
*Wrote a handy dandy little debugging method to visualize both the total
amount of tiles on screen, and the 2d arrays of both grids
*Collision detection for tiles and grids is now squares rather than
circles. But seriously mad props to Jen for writing circle collision
detection that avoided square roots like a motherfucking boss
*Both pre-existing tiles and new ones can be picked up and moved around
~Plans for immediate future~
-Go back over this when It's not 6am and clean it up a little
-Abstract some of the methods I wrote into helper classes to handle the
logic of dropping tiles onto grids and such
-Add occulus rift support and microtransactions

Text if if anything is borked. I'll be awake eventually.