Moving stuff
Not running yet but i wanted to commit what I had~

I added in some stuff to do when commands are interpreted. Still some
implementation to be done when there are "grab" or "function" commands.

I figure once we get the robot moving in tandem with the commands, then
we can add in a timer to slow it down so the sequence of commands is
obvious to the player.

As for moving and turning the code I threw in should do some good
things.

BUT the filtered list of commands thing isnt working, I'm sure we can
figure it out in no time but I'm in a rush so I've linked to a helpful
page and I'll play with it more when I get back to Rochester.

P.S. I added in a gitignore that SHOULD make it so it stops complaining
about our binary files now.