Deleted old file, Button functionality changes
-Found an unused old menu file that wasn't connected to anything and
didn't have a filetype. Looked like an old prototype to our GameScreen
class. So I killed it.

- Added a stop button that takes the place of the go button while the
robot is running. When clicked it returns the robot to its original
position and stops the program from running, This is the same
functionality as the "clear" button but it does it without deleting all
the commands. Though I think we should keep letting clear break
execution and reset the bot the way we have been.

-Added a slight delay in each of the robots steps. Makes it easier to
understand what it's doing in accordance with its commands

-When the robot is finished following its commands the "go" button will
return (it's a "stop" button while the robot is running) and clicking it
will reset the robot and make it run.

-Added a gameState variable to help keep track of when the robot is
running, resetting, or sitting still.

-Gave the robot a "zeroed" property to tell when it has been reset. I
was going to just check if it was on the same square as where it started
instead, but that's not good enough because the user could design a
program to end the robot on the same block as it started :/

TL;DR - Run the robot through a bunch of test simulations, it should
pause nicely each step and not go outside the grid. Try making it stop
while running, after it's finished running, or clearing at any point.
Everything SHOULD be smooth and work swimmingly.