Equation manager
Equations and manager are now abstracted in a code-coherent manner. Data
values are easily accessible and located in sensible places.

A random selection of a pre-made equation can be picked and drawn to the
screen. All operands (the values being added or subtracted to each
other) have accessible board-coordinates, so checking if the bot is
occupying a numbered grid cell should be easy.

NOTE: the board-coordinates are NOT zero-indexed. e.i., (1,1) is the
upper-left hand cell. This just kind of happened when positioning was
taken into account. It could be fixed if needed, but it made sense at
the time.